update version ignored
The db version, display name, and size parameter values are not supported and will be ignored.
This plugin will not work before the callback for the "deviceready" event has been fired, as described in Usage. (This is consistent with the other Cordova plugins.)
The Android version cannot work with more than 100 open db files (due to the threading model used).
UNICODE line separator (\u2028) is and known to be broken in iOS version.
Blob type is currently not supported and known to be broken on multiple platforms.
UNICODE \u0000 (same as \0) character not working in Windows (8.1) (or Windows Phone 8.1) version(s)
iOS version uses a thread pool but with only one thread working at a time due to "synchronized" database access
Large query result can be slow, also due to JSON implementation
ATTACH another database file is not supported (due to path specifications, which work differently depending on the target platform)
User-defined savepoints are not supported and not expected to be compatible with the transaction locking mechanism used by this plugin. In addition, the use of BEGIN/COMMIT/ROLLBACK statements is not supported.